+++ movies/src/main/java/FilmsProject/Interfaces/FilmAccessService.java
@@ -10,6 +10,6 @@ public interface FilmAccessService {
     List<Film> getFilmsByProperty(String property, String value);
     List<Review> getFilmReviews(String filmIdentifier);
     boolean addNewReview(String filmIdentifier,Review review);
-    boolean deleteReview(int reviewId);
-    boolean updateReview(int reviewId, LocalDate date, String reviewText, double rating);
+    boolean deleteReview(Long reviewId);
+    boolean updateReview(Long reviewId, LocalDate date, String reviewText, double rating);
 }
