+++ movies/src/main/java/FilmsProject/Interfaces/UserService.java
@@ -6,5 +6,5 @@ public interface UserService {
     User signIn(String login, String password);
     boolean signUp(String userName, String userLogin, String userPassword);
     boolean writeReview(User author, String filmIdentifier,String reviewText,double rating);
-    boolean updateReview(User authorOfCurrentReview, String filmidentifier, int currentReviewId,String reviewText,double rating);
+    boolean updateReview(User authorOfCurrentReview, String filmidentifier, Long currentReviewId,String reviewText,double rating);
 }
